[🔨refactor] sell, buy 페이지 작업 내용 products 라우터 안으로 이동
